продолжаем

+++ src/main/java/ru/egar/myOrg/worker/repository/WorkHistoryRepository.java
@@ -3,5 +3,5 @@ package ru.egar.myOrg.worker.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.egar.myOrg.worker.model.WorkHistory;
 
-public interface WorkHistoryRepository extends JpaRepository<WorkHistory, Integer> {
+public interface WorkHistoryRepository extends JpaRepository<WorkHistory, Long> {
 }
